refactor(trusty-common): split linear/backend.rs tests out to tests.rs under 500-SLOC cap (#607)
Extracts the #[cfg(test)] mod tests block from backend.rs into a sibling
tests.rs file, which qualifies for the 1500-SLOC test-file budget.
This brings backend.rs from 512 SLOC to 482 SLOC (under the 500 cap).
